// src/Model/Parent.model.ts
export class ParentModel {
    id: string;
    first_name: string;
    last_name: string;
    gender: any;
    relation_ship_type: string;
    email: string;
    mobile: string;
    address: string;
    password: string;
    country_id: string;
    state_id: string;
    city_id: string;
    zip_code: number;
    user_role_id: string;
}